# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  organizationId: number;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  status: string;
  projectCount: number;
  organizationId: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClientRequest {
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description?: string | null;
  status: string;
  budget?: number | null;
  dueDate?: string | null;
  clientId: number;
  clientName: string;
  organizationId: number;
  taskCount: number;
  completedTaskCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectRequest {
  name: string;
  description?: string | null;
  status: string;
  budget?: number | null;
  dueDate?: string | null;
  clientId: number;
}

export interface Phase {
  id: number;
  name: string;
  description?: string | null;
  order: number;
  status: string;
  projectId: number;
  createdAt: string;
  updatedAt: string;
}

export interface Milestone {
  id: number;
  name: string;
  description?: string | null;
  dueDate?: string | null;
  completed: boolean;
  projectId: number;
  phaseId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: string;
  priority: string;
  dueDate?: string | null;
  projectId: number;
  projectName: string;
  assigneeId?: number | null;
  assigneeName?: string | null;
  organizationId: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectDetail {
  id: number;
  name: string;
  description?: string | null;
  status: string;
  budget?: number | null;
  dueDate?: string | null;
  clientId: number;
  clientName: string;
  organizationId: number;
  taskCount: number;
  completedTaskCount: number;
  phases: Phase[];
  milestones: Milestone[];
  tasks: Task[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateTaskRequest {
  title: string;
  description?: string | null;
  status: string;
  priority: string;
  dueDate?: string | null;
  projectId: number;
  assigneeId?: number | null;
}

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  status?: string;
  priority?: string;
  dueDate?: string | null;
  assigneeId?: number | null;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  status: string;
  amount: number;
  dueDate?: string | null;
  paidDate?: string | null;
  clientId: number;
  clientName: string;
  projectId?: number | null;
  projectName?: string | null;
  organizationId: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateInvoiceRequest {
  invoiceNumber: string;
  status: string;
  amount: number;
  dueDate?: string | null;
  clientId: number;
  projectId?: number | null;
}

export interface UpdateInvoiceRequest {
  status?: string;
  amount?: number;
  dueDate?: string | null;
  paidDate?: string | null;
}

export interface CreatePhaseRequest {
  name: string;
  description?: string | null;
  order: number;
  status: string;
  projectId: number;
}

export interface CreateMilestoneRequest {
  name: string;
  description?: string | null;
  dueDate?: string | null;
  completed: boolean;
  projectId: number;
  phaseId?: number | null;
}

export interface ActivityLog {
  id: number;
  action: string;
  entityType: string;
  entityId?: number | null;
  description: string;
  userId?: number | null;
  userName?: string | null;
  organizationId: number;
  createdAt: string;
}

export interface ActivityPage {
  data: ActivityLog[];
  total: number;
  page: number;
  limit: number;
}

export interface RevenueMonth {
  month: string;
  revenue: number;
}

export interface DashboardMetrics {
  totalRevenue: number;
  activeProjects: number;
  totalClients: number;
  overdueInvoices: number;
  overdueAmount: number;
  pendingTasks: number;
  revenueByMonth: RevenueMonth[];
  recentActivity: ActivityLog[];
}

export type GetTasksParams = {
  projectId?: number;
  status?: string;
};

export type GetPhasesParams = {
  projectId?: number;
};

export type GetMilestonesParams = {
  projectId?: number;
};

export type GetActivityParams = {
  page?: number;
  limit?: number;
};