# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 103 Agency OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List all invoices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /phases:
    get:
      operationId: getPhases
      tags: [phases]
      summary: List phases
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of phases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phase"
    post:
      operationId: createPhase
      tags: [phases]
      summary: Create a phase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhaseRequest"
      responses:
        "201":
          description: Created phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
  /milestones:
    get:
      operationId: getMilestones
      tags: [milestones]
      summary: List milestones
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a milestone
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneRequest"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        organizationId:
          type: integer
        createdAt:
          type: string
      required: [id, name, email, role, organizationId, createdAt]
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        status:
          type: string
        projectCount:
          type: integer
        organizationId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, status, projectCount, organizationId, createdAt, updatedAt]
    CreateClientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        status:
          type: string
      required: [name, email, status]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        status:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          nullable: true
        paidDate:
          type: string
          nullable: true
        clientId:
          type: integer
        clientName:
          type: string
        projectId:
          type: integer
          nullable: true
        projectName:
          type: string
          nullable: true
        organizationId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, invoiceNumber, status, amount, clientId, clientName, organizationId, createdAt, updatedAt]
    CreateInvoiceRequest:
      type: object
      properties:
        invoiceNumber:
          type: string
        status:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          nullable: true
        clientId:
          type: integer
        projectId:
          type: integer
          nullable: true
      required: [invoiceNumber, status, amount, clientId]
    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          nullable: true
        paidDate:
          type: string
          nullable: true
    Phase:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
        status:
          type: string
        projectId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, order, status, projectId, createdAt, updatedAt]
    CreatePhaseRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
        status:
          type: string
        projectId:
          type: integer
      required: [name, order, status, projectId]
    Milestone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        completed:
          type: boolean
        projectId:
          type: integer
        phaseId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, completed, projectId, createdAt, updatedAt]
    CreateMilestoneRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        completed:
          type: boolean
        projectId:
          type: integer
        phaseId:
          type: integer
          nullable: true
      required: [name, completed, projectId]
...