# lib/api-zod/src/generated/types/createMilestoneRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMilestoneRequest {
  name: string;
  description?: string | null;
  dueDate?: string | null;
  completed: boolean;
  projectId: number;
  phaseId?: number | null;
}