# lib/api-zod/src/generated/types/createPhaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePhaseRequest {
  name: string;
  description?: string | null;
  order: number;
  status: string;
  projectId: number;
}