# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Invoice {
  id: number;
  invoiceNumber: string;
  status: string;
  amount: number;
  dueDate?: string | null;
  paidDate?: string | null;
  clientId: number;
  clientName: string;
  projectId?: number | null;
  projectName?: string | null;
  organizationId: number;
  createdAt: string;
  updatedAt: string;
}