# lib/api-zod/src/generated/types/updateInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 103 Agency OS API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInvoiceRequest {
  status?: string;
  amount?: number;
  dueDate?: string | null;
  paidDate?: string | null;
}